// frontend/src/app/pages/dashboard/models/TokenModel.ts
export interface TokenModel {

    _id:any;

    id: any;

    tokenName: string;

    tokenSymbol: string;

    tokenTotalSupply:string;

    tokenPrice: string;

    minInvest: string;

    maxInvest: string;

    companyAccount:string;

    duration:number;

    RedemptionType:string;

    yieldValue: string;

    actionType: string;

    amount:string;
    investorCount:number;

    readyToStart:number;

    motifRejection:string;

    status:string;
 

    }